requirements.txt
chromadb
numpy
scipy
google-genai
python-dotenv
ollama